refactor: use argparse
