+++ package.json
@@ -46,24 +46,28 @@
   "homepage": "https://github.com/wechaty/cli#readme",
   "dependencies": {
     "@types/blessed": "^0.1.19",
+    "argparse": "^2.0.1",
     "blessed": "^0.1.81",
     "blessed-contrib": "github:chinggg/blessed-contrib",
     "qrcode-terminal": "^0.12.0",
+    "update-notifier": "^5.1.0",
     "wechaty": "^0.62.3"
   },
   "devDependencies": {
-    "tsconfig-paths": "^3.10.1",
-    "wechaty-puppet-wechat": "^0.28.3",
-    "wechaty-puppet-wechat4u": "^0.18.2",
     "@chatie/eslint-config": "^0.12.4",
     "@chatie/git-scripts": "^0.6.2",
     "@chatie/semver": "^0.4.7",
     "@chatie/tsconfig": "^0.16.2",
     "@grpc/grpc-js": "^1.3.6",
+    "@types/argparse": "^2.0.10",
+    "@types/update-notifier": "^5.1.0",
     "@types/uuid": "^8.3.1",
     "pkg-jq": "^0.2.11",
     "shx": "^0.3.3",
-    "tstest": "^0.4.10"
+    "tsconfig-paths": "^3.10.1",
+    "tstest": "^0.4.10",
+    "wechaty-puppet-wechat": "^0.28.3",
+    "wechaty-puppet-wechat4u": "^0.18.2"
   },
   "publishConfig": {
     "access": "public",
