--- src/interfaces.ts
@@ -1,10 +1,0 @@
-export interface TreeNode {
-  name?: string,
-  children?: TreeChildren | ((node: TreeNode) => TreeChildren | Promise<TreeChildren>),
-  childrenContent?: TreeChildren,
-  extended?: boolean,
-  parent?: TreeNode,
-  [custom: string]: any
-}
-
-export type TreeChildren = Record<string, TreeNode>
